requirements.txt
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.4
django-cors-headers==3.1.0
django-filter==2.2.0
django-import-export==1.2.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
Markdown==3.1.1
odfpy==1.4.0
openpyxl==2.6.3
PyJWT==1.7.1
pytz==2019.2
PyYAML==5.1.2
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
